- Added support for multitasking (now remains paused when in the background) - Added GC achievements (all working, the auto reset on launch) - Added the GC leaderboard - FIXED: The total brogguts being displayed correctly (now it shows all EVER collected, not spent) - Added support for saving SpawnerObjects in saved scenes - Up to and including campaign #8 is implemented - Added expanding/fading ring to explosions - FIXED: The homeBaseLocation wasn't being set correctly

+++ Brogguts/OpenGLEngine/GameCenterStructs.h
@@ -60,8 +60,8 @@ typedef struct Rotating_Entity {
 } RotatingEntityPacket;
 
 typedef struct Complex_Entity {
-	int packetType;
-	int objectID;
+	UInt16 packetType;
+	UInt16 objectID;
 	Vector2f position;
 	Vector2f velocity;
 	float rotation;
